backend/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python==9.2.0
sqlalchemy
